refactor: move fields to core folder
