+++ __tests__/components/fields/FieldColor.spec.js
@@ -2,7 +2,7 @@ import { mountFormGenerator, generatePropsSingleField, generateSchemaSingleField
 import { mount, config } from '@vue/test-utils'
 import { describe, it, expect } from 'vitest'
 
-import FieldColor from '@/fields/input/FieldColor.vue'
+import FieldColor from '@/fields/core/FieldColor.vue'
 
 const form = generateSchemaSingleField(
   'testColor',
