+++ __tests__/components/fields/FieldSelect.spec.js
@@ -2,7 +2,7 @@ import { generateSchemaSingleField, generatePropsSingleField, mountFormGenerator
 import { mount, config } from '@vue/test-utils'
 import { describe, it, expect } from 'vitest'
 
-import FieldSelect from '@/fields/input/FieldSelect.vue'
+import FieldSelect from '@/fields/core/FieldSelect.vue'
 
 const form = generateSchemaSingleField(
   'testSelect',
