+++ __tests__/components/fields/FieldText.spec.js
@@ -2,7 +2,7 @@ import { generateSchemaSingleField, generatePropsSingleField, mountFormGenerator
 import { mount, config } from '@vue/test-utils'
 import { describe, it, expect } from 'vitest'
 
-import FieldText from '@/fields/input/FieldText.vue'
+import FieldText from '@/fields/core/FieldText.vue'
 
 const form = generateSchemaSingleField(
   'textTest',
