+++ src/fields/core/FieldButton.vue
@@ -6,7 +6,7 @@
 
 <script setup>
 import { toRefs } from 'vue'
-import { useFieldProps } from '@/composables'
+import { useFieldProps } from '@/composables/index.ts'
 
 const props = defineProps(useFieldProps())
