+++ src/fields/core/FieldPassword.vue
@@ -23,7 +23,7 @@ import {
   useFieldAttributes,
   useFieldProps,
   useFieldEmits
-} from '@/composables'
+} from '@/composables/index.ts'
 
 const mediumRegex = new RegExp('^(((?=.*[a-z])(?=.*[A-Z]))|((?=.*[a-z])(?=.*[0-9]))|((?=.*[A-Z])(?=.*[0-9])))(?=.{6,})')
 const strongRegex = new RegExp('^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.{8,})')
