+++ src/fields/core/FieldSelectNative.vue
@@ -26,7 +26,7 @@ import {
   useFieldAttributes,
   useFieldProps,
   useFormModel
-} from '@/composables'
+} from '@/composables/index.ts'
 
 const props = defineProps(useFieldProps())
 const emits = defineEmits(useFieldEmits())
