+++ src/fields/core/FieldText.vue
@@ -1,15 +1,14 @@
 <template>
   <input
     :id="id"
-    type="number"
+    class="field-input"
+    type="text"
     :name="field.name"
     :required="isRequired"
     :disabled="isDisabled"
     :placeholder="field.placeholder"
+    :autocomplete="field.autocomplete || 'off'"
     :value="currentModelValue"
-    :max="field.max || undefined"
-    :min="field.min || undefined"
-    :step="field.step || 1"
     @input="onFieldValueChanged"
     @blur="onBlur"
   >
@@ -17,21 +16,15 @@
 
 <script setup>
 import { toRefs } from 'vue'
-import {
-  useFormModel,
-  useFieldAttributes,
-  useFieldValidate,
-  useFieldEmits,
-  useFieldProps
-} from '@/composables'
+import { useFormModel, useFieldAttributes, useFieldValidate, useFieldProps, useFieldEmits } from '@/composables/index.ts'
 
-const props = defineProps(useFieldProps())
 const emits = defineEmits(useFieldEmits())
+const props = defineProps(useFieldProps())
 
 const { field, model } = toRefs(props)
 
-const { isDisabled, isRequired } = useFieldAttributes(model.value, field.value)
 const { currentModelValue } = useFormModel(model.value, field.value)
+const { isRequired, isDisabled } = useFieldAttributes(model.value, field.value)
 const { errors, validate } = useFieldValidate(
   model.value,
   field.value,
@@ -52,9 +45,8 @@ const onBlur = () => {
 
 const onFieldValueChanged = ({ target }) => {
   errors.value = []
-  const step = field.value.step ?? 1
-  const isDecimalStep = step.toString().split('.')[1]
-  if (!isDecimalStep) return emits('onInput', parseInt(target.value))
-  emits('onInput', parseFloat(target.value))
+  emits('onInput', target.value)
 }
+
+defineExpose({ errors })
 </script>
