+++ src/fields/core/FieldTextarea.vue
@@ -24,7 +24,7 @@ import {
   useFieldAttributes,
   useFieldProps,
   useFormModel
-} from '@/composables'
+} from '@/composables/index.ts'
 
 
 const props = defineProps(useFieldProps())
